// src/theme/types.ts
export type ThemeName = "dark" | "light-hc";

export interface ThemePalette {
  surfaceBg: string;
  surfaceSubtleBg: string;
  panelBg: string;
  inputBg: string;
  statusBg: string;
  textPrimary: string;
  textMuted: string;
  textInverse: string;
  accentInfo: string;
  accentSuccess: string;
  accentWarning: string;
  accentError: string;
  accentBrand: string;
  cursor: string;
  mask: string;
  menuSelected: string;
  menuNormal: string;
}